Replace FileStatus with FileStatusOFlag
FileStatus contained only the "mutable" status flags, but was also used
in places where we want the immutable ones, too (e.g. fcntl F_GETFL).

The fcntl F_SETFL handler already separately ensures that we don't
update immutable flags.
