+++ src/main/host/descriptor/socket/unix.rs
@@ -1179,7 +1179,7 @@ impl Protocol for ConnOrientedInitial {
                 return (self.into(), Err(Errno::ECONNREFUSED.into()));
             }
             Err(IncomingConnError::QueueFull) => {
-                if common.status.contains(FileStatus::NONBLOCK) {
+                if common.status.contains(FileStatus::O_NONBLOCK) {
                     return (self.into(), Err(Errno::EWOULDBLOCK.into()));
                 }
 
@@ -2135,7 +2135,7 @@ impl UnixSocketCommon {
             return Err(Errno::EINVAL.into());
         }
 
-        if self.status.contains(FileStatus::NONBLOCK) {
+        if self.status.contains(FileStatus::O_NONBLOCK) {
             flags.insert(MsgFlags::MSG_DONTWAIT);
         }
 
@@ -2245,7 +2245,7 @@ impl UnixSocketCommon {
             return Err(Errno::EINVAL.into());
         }
 
-        if self.status.contains(FileStatus::NONBLOCK) {
+        if self.status.contains(FileStatus::O_NONBLOCK) {
             flags.insert(MsgFlags::MSG_DONTWAIT);
         }
